Desktop/post-quantum-blockchain-phr-security/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
pycryptodome==3.19.0
liboqs-python==0.14.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
